+++ crates/ryft-core/src/broadcasting.rs
@@ -2,7 +2,7 @@ use thiserror::Error;
 
 use crate::{
     parameters::{ParameterError, Parameterized},
-    types::data_type::DataTypeError,
+    types::data_types::DataTypeError,
     types::{ArrayType, DataType, Shape, Size},
 };
 
@@ -64,7 +64,7 @@ pub enum BroadcastingError {
 ///
 /// ```rust
 /// # use ryft_core::broadcasting::{Broadcastable, BroadcastingError};
-/// # use ryft_core::types::data_type::DataType::{Boolean, F32, F64};
+/// # use ryft_core::types::data_types::DataType::{Boolean, F32, F64};
 /// # use ryft_core::types::{ArrayType, Shape};
 /// let x = Shape::new(vec![4.into(), 3.into()]);
 /// let y = Shape::new(vec![3.into()]);
@@ -264,7 +264,7 @@ mod tests {
 
     use crate::parameters::{Parameter, ParameterError};
     use crate::types::Shape;
-    use crate::types::data_type::DataType::*;
+    use crate::types::data_types::DataType::*;
 
     use super::*;