+++ crates/ryft-core/src/errors.rs
@@ -1,7 +1,8 @@
 use thiserror::Error;
 
+use crate::broadcasting::BroadcastingError;
 use crate::parameters::ParameterError;
-use crate::types::data_type::DataTypeError;
+use crate::types::data_types::DataTypeError;
 
 /// Represents errors that can occur in `ryft-core`.
 #[derive(Error, Clone, Debug, Eq, PartialEq, Hash)]
@@ -11,4 +12,7 @@ pub enum Error {
 
     #[error(transparent)]
     DataType(#[from] DataTypeError),
+
+    #[error(transparent)]
+    Broadcasting(#[from] BroadcastingError),
 }