+++ crates/ryft/src/lib.rs
@@ -10,6 +10,6 @@ pub use ryft_core::parameters::{
     PathPrefixedParameterIterator, Placeholder,
 };
 pub use ryft_core::types::Type;
-pub use ryft_core::types::array_type::{ArrayType, Shape, Size};
-pub use ryft_core::types::data_type::{DataType, DataTypeError};
+pub use ryft_core::types::array_types::{ArrayType, Shape, Size};
+pub use ryft_core::types::data_types::{DataType, DataTypeError};
 pub use ryft_macros::{Parameter, Parameterized};